move some function out of AbstractEntity and into AbstractConcreteEntity
